@@ -5,6 +5,7 @@ import (
 	"log/slog"
 
 	"github.com/deckhouse/deckhouse/pkg/log"
+	sdkpkg "github.com/deckhouse/module-sdk/pkg"
 	"github.com/hashicorp/go-multierror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -23,10 +24,10 @@ type OperationSpec struct {
 	Name        string        `json:"name,omitempty" yaml:"name,omitempty"`
 	Subresource string        `json:"subresource,omitempty" yaml:"subresource,omitempty"`
 
-	Object     interface{} `json:"object,omitempty" yaml:"object,omitempty"`
-	JQFilter   string      `json:"jqFilter,omitempty" yaml:"jqFilter,omitempty"`
-	MergePatch interface{} `json:"mergePatch,omitempty" yaml:"mergePatch,omitempty"`
-	JSONPatch  interface{} `json:"jsonPatch,omitempty" yaml:"jsonPatch,omitempty"`
+	Object     any    `json:"object,omitempty" yaml:"object,omitempty"`
+	JQFilter   string `json:"jqFilter,omitempty" yaml:"jqFilter,omitempty"`
+	MergePatch any    `json:"mergePatch,omitempty" yaml:"mergePatch,omitempty"`
+	JSONPatch  any    `json:"jsonPatch,omitempty" yaml:"jsonPatch,omitempty"`
 
 	IgnoreMissingObject bool `json:"ignoreMissingObject" yaml:"ignoreMissingObject"`
 	IgnoreHookError     bool `json:"ignoreHookError" yaml:"ignoreHookError"`
@@ -49,25 +50,19 @@ const (
 )
 
 // GetPatchStatusOperationsOnHookError returns list of Patch/Filter operations eligible for execution on Hook Error
-func GetPatchStatusOperationsOnHookError(operations []Operation) []Operation {
-	patchStatusOperations := make([]Operation, 0)
+func GetPatchStatusOperationsOnHookError(operations []sdkpkg.PatchCollectorOperation) []sdkpkg.PatchCollectorOperation {
+	patchStatusOperations := make([]sdkpkg.PatchCollectorOperation, 0)
 	for _, op := range operations {
-		switch operation := op.(type) {
-		case *filterOperation:
-			if operation.subresource == "/status" && operation.ignoreHookError {
-				patchStatusOperations = append(patchStatusOperations, operation)
-			}
-		case *patchOperation:
-			if operation.subresource == "/status" && operation.ignoreHookError {
-				patchStatusOperations = append(patchStatusOperations, operation)
-			}
+		operation, ok := op.(*patchOperation)
+		if ok && operation.subresource == "/status" && operation.ignoreHookError {
+			patchStatusOperations = append(patchStatusOperations, operation)
 		}
 	}
 
 	return patchStatusOperations
 }
 
-func ParseOperations(specBytes []byte) ([]Operation, error) {
+func ParseOperations(specBytes []byte) ([]sdkpkg.PatchCollectorOperation, error) {
 	log.Debug("parsing patcher operations", slog.String("value", string(specBytes)))
 
 	specs, err := unmarshalFromJSONOrYAML(specBytes)
@@ -76,7 +71,7 @@ func ParseOperations(specBytes []byte) ([]Operation, error) {
 	}
 
 	validationErrors := &multierror.Error{}
-	ops := make([]Operation, 0)
+	ops := make([]sdkpkg.PatchCollectorOperation, 0)
 	for _, spec := range specs {
 		err = ValidateOperationSpec(spec, GetSchema("v0"), "")
 		if err != nil {
@@ -89,23 +84,8 @@ func ParseOperations(specBytes []byte) ([]Operation, error) {
 	return ops, validationErrors.ErrorOrNil()
 }
 
-// Operation is a command for ObjectPatcher.
-//
-// There are 4 types of operations:
-//
-// - createOperation to create or update object via Create and Update API calls. Unstructured, map[string]interface{} or runtime.Object is required.
-//
-// - deleteOperation to delete object via Delete API call. deletionPropagation should be set, default is Foregound.
-//
-// - patchOperation to modify object via Patch API call. patchType should be set. patch can be string, []byte or map[string]interface{}
-//
-// - filterOperation to modify object via Get-filter-Update process. filterFunc should be set.
-type Operation interface {
-	Description() string
-}
-
 type createOperation struct {
-	object      interface{}
+	object      any
 	subresource string
 
 	ignoreIfExists bool
@@ -114,6 +94,18 @@ type createOperation struct {
 
 func (op *createOperation) Description() string {
 	return "Create object"
+}
+
+func (op *createOperation) WithSubresource(subresource string) {
+	op.subresource = subresource
+}
+
+func (op *createOperation) WithIgnoreIfExists(ignore bool) {
+	op.ignoreIfExists = ignore
+}
+
+func (op *createOperation) WithUpdateIfExists(update bool) {
+	op.updateIfExists = update
 }
 
 type deleteOperation struct {
@@ -132,6 +124,10 @@ func (op *deleteOperation) Description() string {
 	return fmt.Sprintf("Delete object %s/%s/%s/%s", op.apiVersion, op.kind, op.namespace, op.name)
 }
 
+func (op *deleteOperation) WithSubresource(subresource string) {
+	op.subresource = subresource
+}
+
 type patchOperation struct {
 	// Object coordinates for patch and delete.
 	apiVersion  string
@@ -141,64 +137,53 @@ type patchOperation struct {
 	subresource string
 
 	// Patch options.
-	patchType           types.PatchType
-	patch               interface{}
+	patchType types.PatchType
+	patch     any
+
+	// using only if not JSON or Merge patch type
+	filterFunc func(*unstructured.Unstructured) (*unstructured.Unstructured, error)
+
 	ignoreMissingObject bool
 	ignoreHookError     bool
 }
 
 func (op *patchOperation) Description() string {
-	return fmt.Sprintf("Patch object %s/%s/%s/%s using %s patch", op.apiVersion, op.kind, op.namespace, op.name, op.patchType)
-}
-
-type filterOperation struct {
-	// Object coordinates for patch and delete.
-	apiVersion  string
-	kind        string
-	namespace   string
-	name        string
-	subresource string
-
-	// Patch options.
-	filterFunc          func(*unstructured.Unstructured) (*unstructured.Unstructured, error)
-	ignoreMissingObject bool
-	ignoreHookError     bool
-}
-
-func (op *filterOperation) Description() string {
 	return fmt.Sprintf("Filter object %s/%s/%s/%s", op.apiVersion, op.kind, op.namespace, op.name)
 }
 
-func NewFromOperationSpec(spec OperationSpec) Operation {
+func (op *patchOperation) hasFilterFn() bool {
+	return op.filterFunc != nil
+}
+
+func (op *patchOperation) WithSubresource(subresource string) {
+	op.subresource = subresource
+}
+
+func (op *patchOperation) WithIgnoreMissingObject(ignore bool) {
+	op.ignoreMissingObject = ignore
+}
+
+func (op *patchOperation) WithIgnoreHookError(ignore bool) {
+	op.ignoreHookError = ignore
+}
+
+func NewFromOperationSpec(spec OperationSpec) sdkpkg.PatchCollectorOperation {
 	switch spec.Operation {
 	case Create:
-		return NewCreateOperation(spec.Object,
-			WithSubresource(spec.Subresource))
+		return NewCreateOperation(spec.Object)
 	case CreateIfNotExists:
-		return NewCreateOperation(spec.Object,
-			WithSubresource(spec.Subresource),
-			IgnoreIfExists())
+		return NewCreateIfNotExistsOperation(spec.Object)
 	case CreateOrUpdate:
-		return NewCreateOperation(spec.Object,
-			WithSubresource(spec.Subresource),
-			UpdateIfExists())
+		return NewCreateOrUpdateOperation(spec.Object)
 	case Delete:
-		return NewDeleteOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name,
-			WithSubresource(spec.Subresource))
+		return NewDeleteOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name)
 	case DeleteInBackground:
-		return NewDeleteOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name,
-			WithSubresource(spec.Subresource),
-			InBackground())
+		return NewDeleteInBackgroundOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name)
 	case DeleteNonCascading:
-		return NewDeleteOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name,
-			WithSubresource(spec.Subresource),
-			NonCascading())
+		return NewDeleteNonCascadingOperation(spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name)
 	case JQPatch:
-		return NewFilterPatchOperation(
-			func(u *unstructured.Unstructured) (*unstructured.Unstructured, error) {
-				filter := jq.NewFilter(app.JqLibraryPath)
-				return applyJQPatch(spec.JQFilter, filter, u)
-			},
+		return NewPatchWithJQOperation(
+			spec.JQFilter,
 			spec.ApiVersion, spec.Kind, spec.Namespace, spec.Name,
 			WithSubresource(spec.Subresource),
 			WithIgnoreMissingObject(spec.IgnoreMissingObject),
@@ -224,70 +209,107 @@ func NewFromOperationSpec(spec OperationSpec) Operation {
 	return nil
 }
 
-func NewCreateOperation(obj interface{}, options ...CreateOption) Operation {
+func NewCreateOperation(obj any) sdkpkg.PatchCollectorOperation {
+	return newCreateOperation(Create, obj)
+}
+
+func NewCreateOrUpdateOperation(obj any) sdkpkg.PatchCollectorOperation {
+	return newCreateOperation(CreateOrUpdate, obj)
+}
+
+func NewCreateIfNotExistsOperation(obj any) sdkpkg.PatchCollectorOperation {
+	return newCreateOperation(CreateIfNotExists, obj)
+}
+
+func newCreateOperation(operation OperationType, obj any) sdkpkg.PatchCollectorOperation {
 	op := &createOperation{
 		object: obj,
 	}
-	for _, option := range options {
-		option.applyToCreate(op)
+
+	switch operation {
+	case Create:
+		// pass
+	case CreateOrUpdate:
+		op.updateIfExists = true
+	case CreateIfNotExists:
+		op.ignoreIfExists = true
 	}
+
 	return op
 }
 
-func NewDeleteOperation(apiVersion, kind, namespace, name string, options ...DeleteOption) Operation {
+func NewDeleteOperation(apiVersion string, kind string, namespace string, name string) sdkpkg.PatchCollectorOperation {
+	return newDeleteOperation(metav1.DeletePropagationForeground, apiVersion, kind, namespace, name)
+}
+
+func NewDeleteInBackgroundOperation(apiVersion string, kind string, namespace string, name string) sdkpkg.PatchCollectorOperation {
+	return newDeleteOperation(metav1.DeletePropagationBackground, apiVersion, kind, namespace, name)
+}
+
+func NewDeleteNonCascadingOperation(apiVersion string, kind string, namespace string, name string) sdkpkg.PatchCollectorOperation {
+	return newDeleteOperation(metav1.DeletePropagationOrphan, apiVersion, kind, namespace, name)
+}
+
+func newDeleteOperation(propagation metav1.DeletionPropagation, apiVersion, kind, namespace, name string) sdkpkg.PatchCollectorOperation {
 	op := &deleteOperation{
 		apiVersion:          apiVersion,
 		kind:                kind,
 		namespace:           namespace,
 		name:                name,
-		deletionPropagation: metav1.DeletePropagationForeground,
+		deletionPropagation: propagation,
 	}
-	for _, option := range options {
-		option.applyToDelete(op)
-	}
+
 	return op
 }
 
-func NewMergePatchOperation(mergePatch interface{}, apiVersion, kind, namespace, name string, options ...PatchOption) Operation {
+func NewMergePatchOperation(mergePatch any, apiVersion string, kind string, namespace string, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
+	return newPatchOperation(types.MergePatchType, mergePatch, apiVersion, kind, namespace, name, opts...)
+}
+
+func NewJSONPatchOperation(jsonpatch any, apiVersion string, kind string, namespace string, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
+	return newPatchOperation(types.JSONPatchType, jsonpatch, apiVersion, kind, namespace, name, opts...)
+}
+
+func newPatchOperation(patchType types.PatchType, patch any, apiVersion, kind, namespace, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
 	op := &patchOperation{
 		apiVersion: apiVersion,
 		kind:       kind,
 		namespace:  namespace,
 		name:       name,
-		patch:      mergePatch,
-		patchType:  types.MergePatchType,
+		patch:      patch,
+		patchType:  patchType,
 	}
-	for _, option := range options {
-		option.applyToPatch(op)
+
+	for _, opt := range opts {
+		opt.Apply(op)
 	}
+
 	return op
 }
 
-func NewJSONPatchOperation(jsonPatch interface{}, apiVersion, kind, namespace, name string, options ...PatchOption) Operation {
+func NewPatchWithJQOperation(jqQuery string, apiVersion string, kind string, namespace string, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
+	return newFilterOperation(func(u *unstructured.Unstructured) (*unstructured.Unstructured, error) {
+		filter := jq.NewFilter(app.JqLibraryPath)
+		return applyJQPatch(jqQuery, filter, u)
+	}, apiVersion, kind, namespace, name, opts...)
+}
+
+func NewPatchWithMutatingFuncOperation(fn func(*unstructured.Unstructured) (*unstructured.Unstructured, error), apiVersion string, kind string, namespace string, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
+	return newFilterOperation(fn, apiVersion, kind, namespace, name, opts...)
+}
+
+func newFilterOperation(fn func(*unstructured.Unstructured) (*unstructured.Unstructured, error), apiVersion, kind, namespace, name string, opts ...sdkpkg.PatchCollectorOption) sdkpkg.PatchCollectorOperation {
 	op := &patchOperation{
 		apiVersion: apiVersion,
 		kind:       kind,
 		namespace:  namespace,
 		name:       name,
-		patch:      jsonPatch,
-		patchType:  types.JSONPatchType,
+		filterFunc: fn,
 	}
-	for _, option := range options {
-		option.applyToPatch(op)
-	}
-	return op
-}
 
-func NewFilterPatchOperation(filter func(*unstructured.Unstructured) (*unstructured.Unstructured, error), apiVersion, kind, namespace, name string, options ...FilterOption) Operation {
-	op := &filterOperation{
-		apiVersion: apiVersion,
-		kind:       kind,
-		namespace:  namespace,
-		name:       name,
-		filterFunc: filter,
+	for _, opt := range opts {
+		opt.Apply(op)
 	}
-	for _, option := range options {
-		option.applyToFilter(op)
-	}
+
 	return op
 }
